move protocols to /cmp
