+++ protocols/cmp/sign/errors.go
