+++ protocols/cmp/sign/local_party.go
@@ -3,13 +3,13 @@ package sign
 import (
 	"github.com/taurusgroup/cmp-ecdsa/pkg/math/curve"
 	"github.com/taurusgroup/cmp-ecdsa/pkg/paillier"
-	"github.com/taurusgroup/cmp-ecdsa/protocols/refresh"
+	refresh2 "github.com/taurusgroup/cmp-ecdsa/protocols/cmp/refresh"
 )
 
 // LocalParty is the state we store for a remote party.
 // The messages are embedded to make access to the attributes easier.
 type LocalParty struct {
-	*refresh.Public
+	*refresh2.Public
 
 	DeltaMtA, ChiMtA *MtA
