+++ protocols/cmp/sign/mta.go
@@ -9,7 +9,7 @@ import (
 	"github.com/taurusgroup/cmp-ecdsa/pkg/math/sample"
 	"github.com/taurusgroup/cmp-ecdsa/pkg/paillier"
 	zkaffg "github.com/taurusgroup/cmp-ecdsa/pkg/zk/affg"
-	"github.com/taurusgroup/cmp-ecdsa/protocols/refresh"
+	refresh2 "github.com/taurusgroup/cmp-ecdsa/protocols/cmp/refresh"
 )
 
 // MtA holds the local data for the multiplicative-to-additive share conversion protocol.
@@ -17,7 +17,7 @@ import (
 // The Sender and Receiver are denoted by index i and j respectively, and we take the perspective of party i.
 // i has share aᵢ and j has share bⱼ
 type MtA struct {
-	Sender, Receiver *refresh.Public
+	Sender, Receiver *refresh2.Public
 
 	// Secret is the senders share aᵢ
 	Secret *curve.Scalar
@@ -52,7 +52,7 @@ type MtA struct {
 // - Kⱼ is the encryption of bⱼ sent by j in the previous round
 func NewMtA(ai *curve.Scalar, Ai *curve.Point,
 	Kj *paillier.Ciphertext,
-	sender, receiver *refresh.Public) *MtA {
+	sender, receiver *refresh2.Public) *MtA {
 
 	beta := sample.IntervalLPrime(rand.Reader)
 
@@ -87,7 +87,7 @@ func NewMtA(ai *curve.Scalar, Ai *curve.Point,
 // public parameters are used.
 // This function is specified as to make clear which parameters must be input to zkaffg.
 // h is a hash function initialized with i's ID
-func (mta *MtA) ProofAffG(h *hash.Hash, verifier *refresh.Public) *MtAMessage {
+func (mta *MtA) ProofAffG(h *hash.Hash, verifier *refresh2.Public) *MtAMessage {
 	if verifier == nil {
 		verifier = mta.Receiver
 	}
@@ -120,7 +120,7 @@ func (mta *MtA) ProofAffG(h *hash.Hash, verifier *refresh.Public) *MtAMessage {
 // - Ki is the receiver's public Kᵢ
 func (mta *MtAMessage) VerifyAffG(h *hash.Hash,
 	Aj *curve.Point, Ki *paillier.Ciphertext,
-	sender, receiver, verifier *refresh.Public) bool {
+	sender, receiver, verifier *refresh2.Public) bool {
 	if verifier == nil {
 		verifier = receiver
 	}
