+++ protocols/cmp/sign/result.go
