+++ protocols/cmp/sign/signature.go
