+++ protocols/cmp/sign/signature_test.go
